now you only progress to the next level if you win
