+++ Gloomhaven/backend/models/level.py
@@ -6,3 +6,4 @@ class Level:
     floor_color_map: list[tuple[int,int]]
     wall_color_map: list[tuple[int,int]]
     monster_classes: list[Character]
+    pre_level_text: str
