+++ Gloomhaven/backend/utils/utilities.py
@@ -1,4 +1,5 @@
 from functools import partial
+from enum import Enum, auto
 
 def make_multiply_modifier(multiplier: int, multiplier_text: str) -> tuple:
     def multiply(x, y):
@@ -9,3 +10,10 @@ def make_additive_modifier(modifier_num) -> tuple:
     def add(x, y):
         return x + y
     return (partial(add, modifier_num), f"{modifier_num:+d}")
+
+class GameState(Enum):
+    START = auto()
+    RUNNING = auto()
+    WIN = auto()
+    GAME_OVER = auto()
+    EXHAUSTED = auto()
